# perf: reduce if statement call on `"` and `\` (micro-optimization)