With this PR when a  `"` or `\` is into the string, the if statement for surrogate check is not called. This is a little performance improvment

benchmark with focus on small string
master
```
short string............................................. x 23,456,041 ops/sec ±0.52% (190 runs sampled)
short string with double quote........................... x 13,475,336 ops/sec ±0.55% (189 runs sampled)
```
PR
```
short string............................................. x 23,641,068 ops/sec ±0.57% (191 runs sampled)
short string with double quote........................... x 13,703,071 ops/sec ±0.43% (187 runs sampled)
```

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [x] documentation is changed or added
- [x] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)